+++ lib/view/airway/checklist/intubation_checklist_page.dart
@@ -9,27 +9,8 @@ import '../../../routes.dart';
 import '../../../strings.dart';
 import '../../../style.dart';
 import '../../../widget/containers/intubation_checklist_container.dart';
+import '../../../widget/lists/checklist_card.dart';
 import '../../view_templates/tab_view_template.dart';
-
-/// Maintains the selected state of every checklist item.
-/// Created by the IntubationChecklistPage screen, so that all checklists
-/// are reset when screen is destroyed.
-class ChecklistSelectionProvider<T> extends ChangeNotifier {
-  final Set<T> _checkedState = {};
-
-  bool isChecked(T item) {
-    return _checkedState.contains(item);
-  }
-
-  void setChecked(T item, {bool checked = true}) {
-    if (checked) {
-      _checkedState.add(item);
-    } else {
-      _checkedState.remove(item);
-    }
-    notifyListeners();
-  }
-}
 
 class IntubationChecklistPage extends StatelessWidget {
   List<String> getChecklistTitles() {
