+++ lib/widget/lists/checklist_card.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../style.dart';
 
-import '../../view/airway/checklist/intubation_checklist_page.dart';
 import '../../widget/cards/reusable_card_base.dart';
 import '../../widget/checkbox.dart';
 
@@ -54,5 +53,25 @@ class _ChecklistItemWidget<T> extends State<ChecklistItemWidget<T>> {
         )
       ],
     );
+  }
+}
+
+/// Maintains the selected state of every checklist item.
+/// Created by the parent screen, so that all checklists
+/// are reset when screen is destroyed.
+class ChecklistSelectionProvider<T> extends ChangeNotifier {
+  final Set<T> _checkedState = {};
+
+  bool isChecked(T item) {
+    return _checkedState.contains(item);
+  }
+
+  void setChecked(T item, {bool checked = true}) {
+    if (checked) {
+      _checkedState.add(item);
+    } else {
+      _checkedState.remove(item);
+    }
+    notifyListeners();
   }
 }
